Store data as Vec instead of CSV
As the original codebase was made with the sole purpose of generating
CSV files, many of the functions were passing around a single String
that represented the comma-delimited CSV. This had the obvious issue
that if the control table was to contain a field with a comma (such
as the ID field of the 2XC430-W250), things would break. Also, this
change improves the type-safety and portability of the functions, which
in the future will enable even more formats.